backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.10
python-decouple==3.8
django-cors-headers==4.7.0
django-filter==24.3
django-extensions==3.2.3
Pillow==10.4.0
celery==5.4.0
redis==5.2.0
gunicorn==23.0.0
whitenoise==6.8.2
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1
coverage==7.6.9